fix: remove unnecessary alignment
I had understood the align field incorrectly. I though it should be a number representing
the amount of bytes needed to align the following field. That doesn't make much sense tho.
Instead, I need to push the amount of bytes needed for the next field to be aligned, that
way when iretq is exectued it can pop the 5 fields correctly. As I make sure the allocated
stack is always aligned, there is no need to push any alignment bytes.