fix: scheduler bug
It could happen that current was idleProc and when doing nextPCB() current gets the value
of tail and then it would break. This should (probably... maybe...) work. nextPCB() switches
current for a process in the list, that gets assigned to ogCurrent, and then another nextPCB
goes to the following process in the list and the loop should work from there.